+++ Changelog.md
@@ -1,5 +1,9 @@
 ### Changelog :
 
+0.13.0 : stability : corrected the scraper in order to allow it to continue to use mangafox ho changed domain name ( .me to .la ) + implemented a new param loop_on_todo in order to allow MangaScrap to download a maximum of todo pages in one go without having to put multiple instructions together<br />
+
+###### Warning : the params database is not compatible<br />Please look at the file migration/0.12.x_to_0.13.x.txt<br />
+
 0.12.10 : debug and new instruction : corrected an issue with the "details" instruction and added a "many-details" instruction to output all of the todo database content<br />
 0.12.9 : stability fix : multiple minor bug fixes<br />
 0.12.8 : stability fix and new feature : it is now possible to add a manga using a link that contains page and chapter values + better management of the redirections<br />